# skillhub_CICD_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkillHub API",
        "description": "API REST de la plateforme SkillHub : authentification, formations, modules, inscriptions et messagerie.",
        "contact": {
            "name": "MU202612"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/formations/{id}/inscription": {
            "post": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "S'inscrire à une formation (apprenant uniquement, max 5)",
                "description": "Inscrit l'apprenant connecté à une formation.\n\nÉtapes :\n- Vérifie le JWT.\n- Refuse si pas apprenant (403).\n- 404 si la formation n'existe pas.\n- 409 si l'apprenant est déjà inscrit à cette formation.\n- 400 si l'apprenant a déjà 5 formations en cours (limite métier).\n- Sinon crée l'inscription avec progression à 0 et log l'événement.\n\nRenvoie aussi le nombre de formations restantes que l'apprenant peut\nencore suivre, pratique pour afficher un compteur côté front.",
                "operationId": "2dc9a1cc6dc9d812479c33692e6c009a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Inscription réussie"
                    },
                    "400": {
                        "description": "Limite de 5 formations atteinte"
                    },
                    "403": {
                        "description": "Pas apprenant"
                    },
                    "404": {
                        "description": "Formation introuvable"
                    },
                    "409": {
                        "description": "Déjà inscrit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Se désinscrire d'une formation",
                "description": "Désinscrit l'apprenant connecté d'une formation.\n\nÉtapes :\n- Vérifie le JWT.\n- Refuse si pas apprenant (403).\n- Cherche l'inscription pour le couple (user, formation).\n- 404 si pas trouvée, sinon supprime.",
                "operationId": "6ff523ac2c06be1613131c4b99cdabe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Désinscription réussie"
                    },
                    "403": {
                        "description": "Pas apprenant"
                    },
                    "404": {
                        "description": "Inscription introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/apprenant/formations": {
            "get": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Mes formations en cours",
                "description": "Renvoie toutes les formations suivies par l'apprenant connecté,\navec la formation et son formateur (id, nom, email) chargés.\n\nSert de tableau de bord apprenant.",
                "operationId": "0adc12194362763212e93a24e2423005",
                "responses": {
                    "200": {
                        "description": "Liste des inscriptions de l'apprenant"
                    },
                    "403": {
                        "description": "Pas apprenant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT renvoyé par /api/login. Préfixer avec 'Bearer '.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}